requirements.txt
absl-py==2.3.1
acme==5.2.2
alembic==1.16.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appnope==0.1.4
asttokens==3.0.0
astunparse==1.6.3
beautifulsoup4==4.14.3
blinker==1.9.0
certbot==5.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorlog==6.9.0
comm==0.2.2
ConfigArgParse==1.7.1
configobj==5.0.9
contourpy==1.3.2
cryptography==46.0.3
cycler==0.12.1
debugpy==1.8.15
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.8.0
executing==2.2.0
fastapi==0.135.1
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==25.2.10
fonttools==4.59.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.74.0
h11==0.16.0
h5py==3.14.0
idna==3.10
ipwhois==1.3.0
ipycanvas==0.14.1
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
josepy==2.2.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
keras==3.11.1
kiwisolver==1.4.8
libclang==18.1.1
librt==0.7.3
lorem==0.1.1
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.3.2
mypy==1.19.0
mypy_extensions==1.1.0
mysql-connector-python==9.6.0
namex==0.1.0
nest-asyncio==1.6.0
numpy==1.26.4
opt_einsum==3.4.0
optree==0.17.0
optuna==4.4.0
packaging==25.0
pandas==2.3.1
parsedatetime==2.6
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
prometheus_client==0.23.1
prompt_toolkit==3.0.51
protobuf==4.25.8
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyOpenSSL==25.3.0
pyparsing==3.2.3
pyRFC3339==2.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
requests==2.32.4
rich==14.1.0
scikit-learn==1.7.1
scipy==1.16.0
seaborn==0.13.2
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.43
stack-data==0.6.3
starlette==0.52.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.16.2
tensorflow-metal==1.2.0
termcolor==3.1.0
threadpoolctl==3.6.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.42.0
wcwidth==0.2.13
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.17.2
